backend/requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.1
attrs==21.4.0
bcrypt==3.2.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
distlib==0.3.4
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.76.0
filelock==3.6.0
greenlet==1.1.2
h11==0.13.0
identify==2.5.0
idna==3.3
iniconfig==1.1.1
Jinja2==3.1.2
MarkupSafe==2.1.1
nodeenv==1.6.0
packaging==21.3
passlib==1.7.4
platformdirs==2.5.2
pluggy==1.0.0
pre-commit==2.19.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.9
pytest==7.1.2
python-decouple==3.6
PyYAML==6.0
requests==2.27.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.18.0
toml==0.10.2
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
virtualenv==20.14.1
